+++ source/Irrlicht/CD3D9Driver.h
@@ -202,10 +202,10 @@ namespace video
 			const core::vector3df& end, SColor color = SColor(255,255,255,255));
 
 		virtual void draw3DLineW(const core::vector3df& start,
-			const core::vector3df& end, SColor color = SColor(255,255,255,255),int width = 0);
+			const core::vector3df& end, SColor color = SColor(255,255,255,255), float width = 0.0f);
 
 		virtual void draw3DShapeW(const core::vector3df* vertices,
-								  u32 vertexCount, SColor color = SColor(255, 255, 255, 255), int width = 0, unsigned short pattern = 0xffff);
+								  u32 vertexCount, SColor color = SColor(255, 255, 255, 255), float width = 0.0f, unsigned short pattern = 0xffff);
 
 		//! initialises the Direct3D API
 		bool initDriver(HWND hwnd, bool pureSoftware);
